backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "This is an application to get weather data for a location"
authors = ["Ross <rocebxyz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests-cache = "^1.2.0"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
pytest = "^8.2.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
